Add `board preferences` and `SensorType`
This commit introduces the concept of global preferences that help manage all those configurations that can change at runtime and needs to be stored inside the board's EEPROM.

On of those configuration is the available sensors; to better manage those it's created the enum `SensorType` that represents all the possible sensors mounted on the board. This in used in conjunction with `has_sensor` to establish what sensor needs to be read.