Secret: Add `arduino_secret.h` to manage secrets
This commit adds an header file called `arduino_secrets.h` that contains all the secrets defined as constants. This file is not added to VCS, but the constants are used inside `config.h`.

TODO: Finda a way to avoid typing this file by hand every time we clone this repository.