MAC: Make `wifi_set_mac_address` method private
Settings the MAC address requires the WiFi connection do be disabled, but the Wi
Fi interface to be initialized. With the current wifi implementation we call `WiFi.mode(WIFI_STA) only inside `wifi_connect` so the WiFi interface is never initialized before we try to connect to a network. For this reason having a public method to set the MAC address will result in constant errors.

In the future we could implement the ability to switch between WiFi station and access point, for this we surely need to change the implementation of `wifi.h`; at this time the method could becam public again.