WiFi Backup: Move task init to ensure it works
In this commit we move the WiFi backup task initialization before trying to connect with the regular connection. This way the backup task is surely started even if we cannot connect to the default connection and the board is waiting for a reboot.

In this commit we also changed the jumper pin number to 26, so it works well with the available pins.