AbstractSensor: Add interface `on_init`, `on_measure`
This commit adds two new virtual methods to `AbstractSensor`: `on_init` and `on_measure`; the old `init` and `measure` virtual methods now are implemented directly by the abstract class and act as a public API while the two protected methods acts as an implementation.
This is done to ensure that all the classes will set and check if the sensor is initialized in a constanty way. Now this check is done in the base class for everyone, and the children needs only to specify their logic.