Fix bugs that emerged from previous changes
In this commit we fix some bugs that have emerged from the latest formatting changes and resulted in a failed compilation.
For some reason the code inside `log.h` is mixed with the one of another branch (probably vscode's fault); another bug is a wrong include inside `sensor_helper.h` that caused cyclic dependencies.