Includes: Fix all includes
In this commit we fix all the `include` directives removing the use of the `common.h` header for including all common headers.
Instead we start following Google C++ Style directives stating that we should include everithing we use; for this reason in each class we added every header that we use (even if already defined inside his header).