Log: Change `log.h` from class to functions
Previously there was a single instance of `Logger` class used globally throughout the code base to log messages; this class relied on a `LogLevel` enum to decide which logs to run.
This approach is very clean, but it doesn't make it possible to eliminate calls to all excluded logging functions from the current level leading to a huge waste of resources and execution time.
In this commit, log functions are defined as macros, while log levels are numeric constants; at compile time we are able to remove all those calls that we know are not used.