Fix bug in SGP30 readings
In this commit we fix a bug where the SGP30 sensor would incorreclty compute the average of the readings.
We also changed how we compute the mean by not removing anymore the max and min valus.
The last change is to not compute anymore the absolute humidity and set it in the senor; we think this was not needed in the fist place since the senor is preatty stable even without it.