Migrate `BoardPreference` to `Preference` class
In this commit we migrate the code of the `BoardPreference` class from the old `EEPROM` class to the new `Preference` class.
The Preference class directly implements the ESP32's Non-Volatile-Storage (NVS) using a key-value pair that simplifyes the reading-writing logic.

We also changes the getter and settes of some preferences grouping them together:
- we added the `BoardInfo` class containing the name, room, location
- we added the `SensorOffsets` class containing the offsets of all the sensors.

Note: In this commit we also partially deprecated the ability to change the ESP32 MAC address. This feature will be deprecated in future commits since it's not needed anymore.